% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CharacterizeVariants.R
\name{CharacterizeVariants}
\alias{CharacterizeVariants}
\title{3'UTR variant characterization}
\usage{
CharacterizeVariants(filename, path_to_filename, path_to_package)
}
\arguments{
\item{filename}{The name of the input vcf file; must be tab delim with 8 columns:
chrom, pos, id, ref, alt, qual, filter, info. Chrom column must be either a number (1) or chr# (chr1)}

\item{path_to_filename}{The path to the directory where the filename is located}

\item{path_to_package}{The path to the R package
#' @return file with variant characterization: processed_filename}
}
\description{
3'UTR variant characterization
}
\examples{
CharacterizeVariants('file.vcf', "~/", "~/CharVar/inst/extdata/");
}
