% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CharVar-package.R
\docType{package}
\name{CharVar-package}
\alias{CharVar}
\alias{CharVar-package}
\title{CharVar: Characterize 3' Untranslated Region Variants by Their Putative Regulatory Elements}
\description{
A program to characterize 3' Untranslated Region variants by their potential regulatory elements. Takes as input canonical 3'UTR variants in a variant call file. Gives as output the variant alternatively polyadenylated isoform location; whether the variant is in an enhanced crosslinking immunoprecipitation (eCLIP) peak, microRNA site, or RNA binding protein motif; any nearby ClinVar, expressed quantitative trait loci (eQTL), or genome wide association study (GWAS) variants; and whether the variant is predicted to be an eQTL or GWAS variant.
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
