% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_reqs.R
\name{install_reqs}
\alias{install_reqs}
\title{RBPamp installation, requires miniconda}
\usage{
install_reqs(path_to_package)
}
\arguments{
\item{path_to_package}{The path to the RegVar R package, check .libPaths if you aren't certain, https://www.rdocumentation.org/packages/base/versions/3.6.2/topics/libPaths}
}
\value{
conda environment- 'RBPamp'- with RBPamp installed, and hg38 fasta installed, ~400MB disc space
}
\description{
RBPamp installation, requires miniconda
}
\examples{
install_reqs('/Library/Frameworks/R.framework/Versions/4.0/Resources/library/RegVar');
}
