% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installRBPamp.R, R/install_reqs.R
\name{installRBPamp}
\alias{installRBPamp}
\title{RBPamp installation (requires miniconda)}
\usage{
installRBPamp(path_to_package)

installRBPamp(path_to_package)
}
\arguments{
\item{path_to_package}{The path to the R package}

\item{path_to_filename}{The path to the directory where the filename is located}
}
\value{
conda environment ('RBPamp') with RBPamp installed

conda environment ('RBPamp') with RBPamp installed, hg38 fasta (~400MB)
}
\description{
RBPamp installation (requires miniconda)

RBPamp installation (requires miniconda)
}
\examples{
CharacterizeVariants('file.vcf', "~/", "~/CharVar/inst/extdata/");
CharacterizeVariants('file.vcf', "~/", "~/CharVar/inst/extdata/");
}
