% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CharacterizeVariants_single_input.R
\name{CharacterizeVariants_single_input}
\alias{CharacterizeVariants_single_input}
\title{3'UTR variant characterization}
\usage{
CharacterizeVariants_single_input(path_to_output)
}
\arguments{
\item{path_to_output}{The path to the directory where you would like the output file written}
}
\value{
output file with variant characterization, named processed_chrom_chromEnd_ref_alt
}
\description{
3'UTR variant characterization
}
\examples{
CharacterizeVariants_single_input("/Library/Frameworks/R.framework/Versions/4.0/Resources/library/RegVar", "~/");
}
