% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installRBPamp.R
\name{installRBPamp}
\alias{installRBPamp}
\title{RBPamp installation (requires miniconda)}
\usage{
installRBPamp(path_to_package)
}
\arguments{
\item{path_to_package}{The path to the R package}

\item{path_to_filename}{The path to the directory where the filename is located}
}
\value{
conda environment ('RBPamp') with RBPamp installed
}
\description{
RBPamp installation (requires miniconda)
}
\examples{
CharacterizeVariants('file.vcf', "~/", "~/RegVar/inst/extdata/");
}
